coreplan_agents_assistix/CorePlanAssistix/requirements.txt
flask
openai
pandas
mysql-connector-python
dash
python-dotenv
matplotlib